+++ libc/stdlib/strtof.c
@@ -25,5 +25,5 @@
  */
 
 float strtof(const char *nptr, char **endptr) {
-	return strtold(nptr, endptr);	
+	return (float) strtold(nptr, endptr);	
 }
